+++ lib/glimmer/libui/figure.rb
@@ -27,9 +27,10 @@ module Glimmer
     # Can optionally have `closed true` property (connecting last point to first point automatically)
     class Figure < Shape
       parameters :x, :y
+      parameter_defaults nil, nil
     
       def draw(area_draw_params)
-        ::LibUI.draw_path_new_figure(path_proxy.libui, *@args) unless @args.empty? # TODO if args empty then wait till there is an arc child and it starts the figure
+        ::LibUI.draw_path_new_figure(path_proxy.libui, *@args) unless @args.compact.empty? # TODO if args empty then wait till there is an arc child and it starts the figure
         children.each {|child| child.draw(area_draw_params)}
         ::LibUI.draw_path_close_figure(path_proxy.libui) if closed?
         super